requirements.txt
Django>=4.2
djangorestframework
channels
channels-redis
celery>=5.0
redis
textblob
scikit-learn
beautifulsoup4
lxml
requests
python-dotenv
transformers
sentence-transformers
bertopic[all]
django-cors-headers
torch
kombu>=5.0
daphne
